+++ packages/uring_api/src/_uring_api/uring_api_capi.c
@@ -396,12 +396,14 @@ int UringApiCapi_RingBreakWait(PyObject *ring) {
 
 PyObject *UringApiCapi_RingWait(PyObject *ring, double timeout) {
     struct __kernel_timespec timeout_value;
-    int timeout_kind;
     if (!ring_type_check(ring)) {
         return NULL;
     }
     if (timeout < 0.0) {
-        return UringApiRing_wait_impl((UringApiRing *)ring, 0, NULL, false);
+        return UringApiRing_wait_impl((UringApiRing *)ring, URING_API_WAIT_BLOCKING, NULL, false, NULL);
+    }
+    if (timeout == 0.0) {
+        return UringApiRing_wait_impl((UringApiRing *)ring, URING_API_WAIT_PEEK, NULL, false, NULL);
     }
     timeout_value.tv_sec = (long long)timeout;
     timeout_value.tv_nsec = (long long)((timeout - (double)timeout_value.tv_sec) * 1000000000.0);
@@ -411,8 +413,7 @@ PyObject *UringApiCapi_RingWait(PyObject *ring, double timeout) {
     if (timeout_value.tv_nsec > 999999999) {
         timeout_value.tv_nsec = 999999999;
     }
-    timeout_kind = timeout == 0.0 ? 2 : 1;
-    return UringApiRing_wait_impl((UringApiRing *)ring, timeout_kind, &timeout_value, false);
+    return UringApiRing_wait_impl((UringApiRing *)ring, URING_API_WAIT_TIMEOUT, &timeout_value, false, NULL);
 }
 
 int UringApiCapi_RingSetCallback(PyObject *ring, PyObject *callback) {